--- a/lib/core-integration/cardano-wallet-core-integration.cabal
+++ b/lib/core-integration/cardano-wallet-core-integration.cabal
@@ -77,6 +77,7 @@ library
     , text
     , text-class
     , time
+    , timeit
   hs-source-dirs:
       src
   exposed-modules: