# Issue Number

<!-- Put here a reference to the issue this PR relates to and which requirements it tackles -->
https://jira.iohk.io/browse/ADP-586

# Overview

<!-- Detail in a few bullet points the work accomplished in this PR -->

- [x] I have tried to replicate performance problem with many-addresses-byron-wallet

# Comments

For 100k Unused addresses
```
1. Importing Unused random addresses - 82.63s 
2. Get of one just imported Unused random address - 9.63s
3. Send money to this address - 0.43s
4. Estimate tx fee using byron wallet - 0.01s
5. Submit tx fee using byron wallet - 7.55s
```

For 150k Unused addresses
```
1. Importing Unused random addresses - 243.43s 
2. Get of one just imported Unused random address - 17.61s
3. Send money to this address - 0.41s
4. Estimate tx fee using byron wallet - 0.01s
5. Submit tx fee using byron wallet - 11.25s
```

For 200k Unused addresses
```
1. Importing Unused random addresses - 449.24s 
2. Get of one just imported Unused random address - 21.72s
3. Send money to this address - 0.41s
4. Estimate tx fee using byron wallet - 0.01s
5. Submit tx fee using byron wallet - 22.84s
```

For 150k Unused addresses for tx with 10 inputs and 9 outputs
```
1. Importing Unused random addresses - 250.94s
2. Get of one just imported Unused random address - 13.27s
3. Send money to this address - 2.62s
4. Estimate tx fee using byron wallet - 0.01s
5. Submit tx fee using byron wallet - 14.35s
```
Using Byron random addresses with small or big relative indices does not influence the results.

**Interesting result**
For 100k Unused addresses for tx with 10 inputs and 9 outputs 
```
1. Importing Unused random addresses - 85.52s
2. Get of one just imported Unused random address - 7.15s
3a. Send money to this address - 2.29s
3b. Send money to this address - 2.22s
3c. Send money to this address - 2.50s
3d. Send money to this address - 2.55s
3e. Send money to this address - 2.41s
4. Estimate tx fee using byron wallet - 3.20s
5a. Submit tx fee using byron wallet - 7.39s
5b. Submit tx fee using byron wallet - 6.55s
5c. Submit tx fee using byron wallet - 4.09s
5d. Submit tx fee using byron wallet - 14.26s
5e. Submit tx fee using byron wallet - 97.02s
``` 
5a-5e is just submitting txs one by one, the same for 3a-3e.
the difference is that in 3a-3e case we are sending from fixture small-number-addresses-byron-wallet.
For 5a-5e we are sending from big-number-of-addresses-byron-wallet

<!-- Additional comments or screenshots to attach if any -->

<!-- 
Don't forget to:

 ✓ Self-review your changes to make sure nothing unexpected slipped through
 ✓ Assign yourself to the PR
 ✓ Assign one or several reviewer(s)
 ✓ Once created, link this PR to its corresponding ticket
 ✓ Assign the PR to a corresponding milestone
 ✓ Acknowledge any changes required to the Wiki
-->